:hammer: chore: refactor active map calls
