+++ packages/client/locales/en/calls.json
@@ -37,6 +37,8 @@
     "purge": "Purge",
     "purgeSelected": "Purge selected",
     "purgeSelectedCalls": "Purge selected calls",
+    "toggleCall": "Toggle Call",
+    "removeMarker": "Remove Marker",
     "alert_purgeSelectedCalls": "Are you sure you want to purge {length} item(s)? This will delete {length} item(s) from the database. This action cannot be undone.",
     "alert_deleteCallEvent": "Are you sure you want to delete this event? This action cannot be undone.",
     "alert_end911Call": "Are you sure you want to end this call?",
